refactor(domaininput): renames customtextinput to domaininput and moves it to chronicle folder in components
SXT-1023
